+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
-* 1.1.0 Fixes display of html-only emails.  (Thanks @hugotacito)
+* 1.2.0 Removes jquery dependence.  Compatible with django-debug-toolbar 1.10+.
+* 1.1.0 Fixes display of html-only emails.  (Thanks @hugotacito!)
 * 1.0.0 Initial release